Migrate FileViewer to TypeScript
