+++ src/ui/FileViewer/data.mock.ts
@@ -8,7 +8,7 @@ export const msg0 = {
   updatedAt: 0,
   channelType: "group",
   mentionType: "users",
-  url: "https://sendbird-staging.s3.amazonaws.com/26FA710D-A7F6-4483-ABEA-1F07A07A112E/upload/n/e49e0a4111d4445a8840bd51f3c9d933.png",
+  url: "https://static.sendbird.com/sample/profiles/profile_15_512px.png",
   name: "sent-mail (1).png",
   size: 9250,
   type: "image/png",
@@ -35,7 +35,7 @@ export const msg1 = {
   updatedAt: 0,
   channelType: "group",
   mentionType: "users",
-  url: "https://sendbird-staging.s3.amazonaws.com/26FA710D-A7F6-4483-ABEA-1F07A07A112E/upload/n/768e0a1481d74901bb0ace8229ec253a.mp4",
+  url: "https://static.sendbird.com/sample/profiles/profile_15_512px.png",
   name: "chris-ddack.mp4",
   size: 3651596,
   type: "video/mp4",
